refactor: move robots analyzer and analytics test into tests/ folders
