+++ entrypoints/popup/utils/robots.ts
